Move some utility functions from main.rs to utils.rs
