+++ src/utils.rs
@@ -1,4 +1,6 @@
 use bitcoin::util::hash::Sha256dHash;
+use secp256k1::key::PublicKey;
+use secp256k1::Secp256k1;
 
 pub fn hex_to_vec(hex: &str) -> Option<Vec<u8>> {
 	let mut out = Vec::with_capacity(hex.len() / 2);
@@ -46,3 +48,34 @@ pub fn hex_to_u256_rev(hex: &str) -> Option<Sha256dHash> {
 	Some(Sha256dHash::from_data(&out))
 }
 
+pub fn hex_to_compressed_pubkey(hex: &str) -> Option<PublicKey> {
+	let data = match hex_to_vec(&hex[0..33*2]) {
+		Some(bytes) => bytes,
+		None => return None
+	};
+	match PublicKey::from_slice(&Secp256k1::without_caps(), &data) {
+		Ok(pk) => Some(pk),
+		Err(_) => None,
+	}
+}
+
+#[inline]
+pub fn hex_str(value: &[u8]) -> String {
+	let mut res = String::with_capacity(64);
+	for v in value {
+		res += &format!("{:02x}", v);
+	}
+	res
+}
+
+#[inline]
+pub fn slice_to_be64(v: &[u8]) -> u64 {
+	((v[0] as u64) << 8*7) |
+	((v[1] as u64) << 8*6) |
+	((v[2] as u64) << 8*5) |
+	((v[3] as u64) << 8*4) |
+	((v[4] as u64) << 8*3) |
+	((v[5] as u64) << 8*2) |
+	((v[6] as u64) << 8*1) |
+	((v[7] as u64) << 8*0)
+}
